Refactor package structures for API categorization
Reorganized classes and tests under new `common.api` package structure to better align with API responsibilities. Updated imports and package-info files accordingly to reflect the changes.
