+++ crime-assessment-service/src/main/java/uk/gov/justice/laa/crime/assessmentservice/common/api/filter/WebClientFilters.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.assessmentservice.common.filter;
+package uk.gov.justice.laa.crime.assessmentservice.common.api.filter;
 
 import lombok.experimental.UtilityClass;
 import lombok.extern.slf4j.Slf4j;
