+++ crime-assessment-service/src/main/java/uk/gov/justice/laa/crime/assessmentservice/common/api/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.modulith.NamedInterface("api")
+package uk.gov.justice.laa.crime.assessmentservice.common.api;
