--- crime-assessment-service/src/main/java/uk/gov/justice/laa/crime/assessmentservice/common/package-info.java
@@ -1,1 +1,0 @@
-package uk.gov.justice.laa.crime.assessmentservice.common;
