+++ crime-assessment-service/src/test/java/uk/gov/justice/laa/crime/assessmentservice/filter/Resilience4jRetryFilterTest.java
@@ -7,7 +7,7 @@ import io.github.resilience4j.retry.RetryConfig;
 import io.github.resilience4j.retry.RetryRegistry;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
-import uk.gov.justice.laa.crime.assessmentservice.common.filter.Resilience4jRetryFilter;
+import uk.gov.justice.laa.crime.assessmentservice.common.api.filter.Resilience4jRetryFilter;
 
 import java.net.URI;
 import java.util.Arrays;
